added matura counter

+++ www/cytaty.php
@@ -1,7 +1,6 @@
 <?php
 session_start();
-if(!isset($_SESSION['loggedin'])&&$_SESSION['loggedin']!=true)
-{
+if (!isset($_SESSION['loggedin']) && $_SESSION['loggedin'] != true) {
 	header("Location: index.php");
 	die();
 }
@@ -10,7 +9,12 @@ require_once("dbutils.php");
 include("requestuserdata.php");
 
 # check for quotes waiting to be removed
-if(!empty($_POST['postToRemove'])){
+if (!empty($_POST['postToRemove'])) {
+	if(($_SESSION['redaktor']==0)||($_SESSION['userisadmin']==0))
+	{
+		header("Location: main.php");
+		die();
+	}
 	$postID = $_POST['postToRemove'];
 	$query = "DELETE FROM cytaty WHERE id=?";
 	easyQuery($query, "i", $postID);
@@ -27,18 +31,18 @@ if(!empty($_POST['postToRemove'])){
 <!DOCTYPE html>
 <html>
 <head>
-  <?php require('standardHead.php'); ?>
-  <link rel="stylesheet" type="text/css" href="quotestyles.css">
+	<?php require('standardHead.php'); ?>
+	<link rel="stylesheet" type="text/css" href="quotestyles.css">
 	<link rel="stylesheet" type="text/css" href="quotes.css">
 	<link rel="stylesheet" type="text/css" href="divColors.css">
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
 	<script src="quotelikes.js"></script>
 	<script src="quotes.js"></script>
 </head>
 <body>
-  <div class="container">
-    <?php require('topbar.php') ?>
-    <div class="content">
+	<div class="container">
+		<?php require('topbar.php') ?>
+		<div class="content">
 			<div class="coverBar">
 				<div class="optionBar">
 					<div class="left">
@@ -56,19 +60,19 @@ if(!empty($_POST['postToRemove'])){
 			</div>
 			<div class="alertBox">
 				<?php
-					if(isset($_SESSION['quoteModification'])) {
-						if($_SESSION['quoteModification'] == "add") {
-							echo "Pomyślnie dodano nowy cytat!<br>";
-						}
-						if($_SESSION['quoteModification'] == "remove") {
-							echo "Pomyślnie usunięto cytat!<br>";
-						}
-						unset($_SESSION['quoteModification']);
+				if (isset($_SESSION['quoteModification'])) {
+					if ($_SESSION['quoteModification'] == "add") {
+						echo "Pomyślnie dodano nowy cytat!<br>";
 					}
+					if ($_SESSION['quoteModification'] == "remove") {
+						echo "Pomyślnie usunięto cytat!<br>";
+					}
+					unset($_SESSION['quoteModification']);
+				}
 				?>
-		</div>
+			</div>
 			<div id=quoteBox> <?php require('loadquotes.php'); ?> </div>
 		</div>
-  </div>
+	</div>
 </body>
 </html>
